change the tools to a module
