+++ Tools/src/main/scala/tools/util/Command.scala
@@ -1,0 +1,6 @@
+package de.htwg.se.uno.util
+
+trait Command:
+  def doStep(): Unit
+  def undoStep(): Unit
+  def redoStep(): Unit
