+++ Tools/src/main/scala/tools/util/UndoManager.scala
